requirements.txt
onnxruntime-gpu
tritonclient[http]
numpy
onnx 
torch
torchvision